// packages/memory-monitor/src/types.ts
export interface TriggerPosition {
  top?: number | string;
  bottom?: number | string;
  right?: number | string;
  left?: number | string;
}